% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berm.R
\name{ElasticNet.model}
\alias{ElasticNet.model}
\title{Elastic Net Model}
\usage{
ElasticNet.model(
  x,
  y,
  lambda_grid = 10^seq(2, -3, by = -0.1),
  alpha_grid = seq(1, 0, length.out = 20),
  standardize = TRUE,
  nfold = 10
)
}
\arguments{
\item{x}{A matrix or data frame of predictor variables where rows are observations
and columns are variables.}

\item{y}{A vector of the response variable.}

\item{lambda_grid}{A numeric vector of lambda values to be tested.}

\item{alpha_grid}{A numeric vector of alpha values to be tested.}

\item{standardize}{A logical value indicating whether the predictor variables should
be standardized before fitting the model.}

\item{nfold}{The number of folds to be used in cross-validation. This parameter
determines how the data is split during the cross-validation process.}
}
\value{
A list containing several components:
- 'cvres': A data frame containing the cross-validation results.
- 'pars': The best tuning parameters (lambda and alpha) selected based on
cross-validation.
- 'coef': A data frame of coefficient estimates at the best lambda and alpha.
- 'intercept': The intercept term from the model at the best lambda.
- 'cvmodels': The full cross-validated \code{train} object including all models
fitted during the tuning process.
- 'finalmodel': The final model object with parameters set at the best lambda
and alpha.
}
\description{
This function fits an elastic net model to the provided data using cross-validation
to select the best lambda and alpha values.
}
