% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluations.R
\name{eval_results}
\alias{eval_results}
\title{Evaluate Model Results}
\usage{
eval_results(true, predicted)
}
\arguments{
\item{true}{A numeric vector of observed true values of the dependent variable.}

\item{predicted}{A numeric vector of predicted values from the model, corresponding
to the true values.}
}
\value{
A single numeric value representing the R-squared of the model.
}
\description{
This function calculates the R-squared value. The R^2 value provides an indication
of goodness of fit and therefore a measure of how well unseen samples are likely
to be predicted by the model, with values ranging from 0 to 1.
}
