% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berm.R
\name{predict.berm}
\alias{predict.berm}
\title{Prediction with BERM}
\usage{
\method{predict}{berm}(object, newdata)
}
\arguments{
\item{object}{The \code{berm} model object containing \code{finalmodel}, \code{coef}, and \code{intercept}.}

\item{newdata}{A data frame or matrix containing new observations. Columns must
match those used in the model training.}
}
\value{
Returns a vector of predicted values based on \code{newdata}.
}
\description{
Predicts new values using the model obtained from \code{berm}.
This function computes the predictions.
}
