% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berm.R
\name{summary.berm}
\alias{summary.berm}
\title{Summary of BERM}
\usage{
\method{summary}{berm}(object)
}
\arguments{
\item{object}{The \code{berm} model object containing at least \code{intercept} and \code{coef}
components.}
}
\value{
The function prints the model coefficients.
}
\description{
This function displays the final model coefficients from a berm model object,
including the intercept and predictors' coefficients. It provides a quick overview
of the model's parameters.
}
